Added an example swdiag-server for running on Unix systems. Only tested on Linux (redhat) as yet. This is still a work in progress. But am checkpointing a commit to the repo "just in case".
There is a daemon that reads a configuration, and then executes modules to get their test/rule/action configuration. It also runs up a web server for accessing that information, and sends emails if required. Essentially this commit contains the majority of the infrastructure, but needs a lot of polish to make it usable.